24-correcion-de-errores-parte-visual - FIX: se arreglo una serie de problemas visuales al correr la pagina, por la distribución en pages
